Fixed & Completed Basic Project-2
Project-2 was never finished due to issues I wasn't able to resolve so was never uploaded to the server. However, I have revisited this project and completed a functioning application. The main purpose of this application is to demonstrate how Carbon is used within Laravel. I plan to add some more features:
- A way to filter the tasks by their due/created date instead of just showing them by earliest due date.
- A user authentication system.
- A task assignment system to assign tasks to users, possibly within a group.